Added transfer operator and AVX512 for QPhix and Coarse VCycles The QPhiX->Coarse operations work well for a 16^4 fine 4^4 coarse
However Coarse->Coarse could probably benefit from some kind of nested
parallelism (and this could be fed back I guess to the QPhiX->Coarse Op
as needed).

E.g. In restriction the sum over the sites within the block could be parallelized
and care must then be taken with the sum part. In prolongation again one could imagine
having the open MP loop over the fine sites... (right now it is over the coarse ones).
but also one would want to potentially parallelize over the colorspins of the output
but that should be done at the level of cachelines. Ideally multiple threads even
in the same core ought not to write the same cacheline simulataneously