Major refactor of the CoarseGauge type. The 'clover term' is now not going to be the 9-th direction, because that would mean the inverse would be duplicated in both the DA and the AD terms. So the new refactoring has clover diagonal, inv_diagonal, off_diagonal, ad and da fields and accessors.
A whole bunch of incidental and induced changes came to ensure the tests pass.
E.g. ZeroGauge needs to zero all the parts. Several places just looped over 9
directions, or touched the 9th direction in a special way. These needed to be
changed.

As of this commit, all the tests pass.