Added Even Odd Structure. Tested VCycle Application reduces preconditioner residuum when not doing all the wrapping. This makes EO a lot faster. Further optimization is to optimize the restrictors (knowing that the input vector is zero on even checkerboard) and the prolongators knowing that the output vector will be ignored on the output checkerboard.
NB: It is always the unprec operator whose nullspace is used. We can use an even-odd
wrapper for this, but we mustn't use the EvenOdd Operator as in that case when coarsening
Dslash with the unprec operator (DslashDir) it will pull in only zeros from the neighbors
and essentially only the clover term will get coarsened.  This can lead to issues downstream
(e.g. I had it affect the null space solve for the lower level, which had a BiCGStab Breakdown
-- somehow, through a convolution of scaling factors of a clover term between the operators,
the coarse op matrix turned out to be the identity, and this caused a breakdown (t == 0).